+++ inventario/views/__init__.py
@@ -1,3 +1,4 @@
 
 from .maestrocintas import login, DetalleProgramasDeleteView, DetalleProgramasUpdateView, DetalleProgramasCreateView, DetalleProgramasDetailView, DetalleProgramasCreateView, DetalleProgramasUpdateView, DetalleProgramasDetailView, DetalleProgramasCreateView, DetalleProgramasListView, MaestroCintasFormView, MaestroCintasDeleteView, MaestroCintasUpdateView, MaestroCintasDetailView, MaestroCintasCreateView, MaestroCintasListView, FormatosCintasCreateView, FormatosCintasDetailView, FormatosCintasListView, FormatosCintasFormView
-from .prestamos import PrestamosListView, GetFolioPrestamo, GetFolioDetail, RegisterInVideoteca, ValidateOutVideoteca, RegisterOutVideoteca,PrestamoDetalle,Filtrar_prestamos, generar_pdf, generar_pdf_modal
+from .prestamos import PrestamosListView, GetFolioPrestamo, GetFolioDetail, RegisterInVideoteca, ValidateOutVideoteca, RegisterOutVideoteca,PrestamoDetalle, Filtrar_prestamos
+from .reports import generar_pdf, generar_pdf_modal
