+++ inventario/views/reports.py
@@ -2,90 +2,202 @@ import textwrap, operator, base64, json, datetime
 from django.conf import settings
 from django.views.decorators.csrf import csrf_exempt
 from pyreportjasper import PyReportJasper
+from django.core import serializers
+from fpdf import FPDF
+from io import BytesIO
+from reportlab.pdfgen import canvas
+from reportlab.lib.pagesizes import letter
+from reportlab.lib.units import inch
+from ..models import Prestamos, DetallePrestamos, MaestroCintas, DetalleProgramas, Videos
+from django.http.response import HttpResponse, JsonResponse
 
-@csrf_exempt      
-def json_to_pdf(request,fechaInicio,fechaFin,folio  ):
-    dateInicio = datetime.datetime.strptime(fechaInicio, "%d-%m-%Y")
-    dateFin = datetime.datetime.strptime(fechaFin, "%d-%m-%Y")
- 
+class PDF(FPDF):
+    def __init__(self, orientation='P', unit='mm', format='A4', q=None):
+        super().__init__(orientation, unit, format)
+        self.q = q
+
+    def header(self):
+        
+        # Configuración de la cabecera del PDF
+        self.image('images/EducaciónAprende.jpeg', x=10, y=8, w=50)
+        self.image('images/logo-aprendemx.png', x=65, y=5, w=50)
+        self.ln()
+
+        self.set_font('Arial', 'B', 8)
+        self.cell(480,1, 'SECRETARÍA DE EDUCACIÓN PÚBLICA', 0, 10, 'C')
+        self.ln(3)
+        self.cell(440,1, 'Subdirección de Sistematización de Acervos y Desarrollo Audiovisual', 0, 20, 'C')
+        self.ln(3)
+        self.cell(518,1, 'Audiovisual', 0, 20, 'C')
+        self.ln(3)
+        self.cell(458,1, 'Departamento de Conservación de Acervos Videográficos', 0, 20, 'C')
+        self.ln(40)
+
+        self.set_font('Arial', 'B', 8)
+        self.cell(84, 10,   'NOMBRE:      _______________________________', 0, 0, 'C')
+        # self.ln(7)
+        self.cell(306, 10,  'DIRECCIÓN:   _______________________________', 0, 0, 'C')
+        self.ln(17)
+        self.cell(83, 10,   'PUESTO:      _______________________________', 0, 0, 'C')
+
+        self.cell(103, 10,  'EXTENSIÓN:   _______________________________', 0, 0, 'C')
+        self.cell(103, 10,  'CORREO:      _______________________________', 0, 0, 'C')
+        self.ln(40)
+        
+        self.set_font('Arial', 'B', 15)
+        self.cell(280, 10, f'Prestamos de la cinta ({self.q})', 0, 0, 'C')
+        self.ln(20)
+        
+    def footer(self):
+
+        self.ln(40)
+        self.set_font('Arial', 'B', 8)
+        self.cell(103, 10, 'RECIBE:', 0, 0, 'C')
+        self.cell(200, 10, 'DEVUELVE:', 0, 0, 'C')
+        self.ln()
+        #self.drawline    
+        self.cell(103, 10, '________________________________________', 0, 0, 'C')
+        # self.ln(8)
+        self.cell(200, 10, '________________________________________', 0, 0, 'C')
+        # self.ln()
+
+        # Configuración del pie de página del PDF
+        self.set_y(-15)
+        self.set_font('Arial', 'I', 8)
+        self.cell(0, 10, 'Página %s' % self.page_no(), 0, 0, 'C')
+
+    def generate_table(self, data):
+        # Generación del contenido de la tabla en el PDF
+        self.set_fill_color(144, 12, 63)
+        self.set_text_color(255, 255, 255) # Establece el color de la letra en blanco
+        self.cell(40, 10, 'Folio', 1, 0, '', True)
+        self.cell(40, 10, 'Usuario', 1, 0, '', True)
+        self.cell(80, 10, 'Fecha y Hora Prestamo', 1, 0, '', True)
+        self.cell(80, 10, 'Fecha de devolución', 1, 0, '', True)
+        self.cell(40, 10, 'Estatus', 1, 0, '', True)
+        self.set_text_color(0, 0, 0)
+        self.ln()
+
+        for row in data:
+            self.cell(40, 10, str(row['pres_folio']), 1)
+            self.cell(40, 10, str(row['usua_clave']), 1)
+            self.cell(80, 10, str(row['pres_fechahora']), 1)
+            self.cell(80, 10, str(row['pres_fecha_devolucion']), 1)
+            self.cell(40, 10, str(row['pres_estatus']), 1)
+            self.ln()
+
+def generar_pdf(request):
+    q = request.GET.get('q')
+    detalle_prestamos = DetallePrestamos.objects.filter(vide_codigo=q)
+    pres_folios = detalle_prestamos.values_list('pres_folio_id', flat=True)
+    prestamos_data = []
+    for pres_folio_id in pres_folios:
+        prestamo = Prestamos.objects.filter(pres_folio=pres_folio_id).first()
+        if prestamo:
+            prestamo_data = {
+                "pres_folio": prestamo.pres_folio,
+                "usua_clave": prestamo.usua_clave,
+                "pres_fechahora": prestamo.pres_fechahora,
+                "pres_fecha_devolucion": prestamo.pres_fecha_devolucion,
+                "pres_estatus": prestamo.pres_estatus
+            }
+            prestamos_data.append(prestamo_data)
+
+    response = HttpResponse(content_type='application/pdf')
+    # response['Content-Disposition'] = 'attachment; filename="Videoteca.pdf"'
+    response['Content-Disposition'] = f'attachment; filename="Videoteca_Código_{q}.pdf"'
+    pdf = PDF('P', 'mm', (300, 350), q)
+    # Abre una nueva página en el documento PDF
+    pdf.add_page()
+
+    # Agrega los datos de los préstamos a la página actual
+    pdf.generate_table(prestamos_data)
+    response.write(pdf.output(dest='S').encode('latin1'))
+    return response
+# ---------------------------------PDF2----------------------------------------------------------------------------------------
+# @csrf_exempt    
+class GENERATE(FPDF):
+    def __init__(self, orientation='P', unit='mm', format='A4', q=None):
+        super().__init__(orientation, unit, format)
+        self.q = q
+
+    def header(self):
+        
+        # Configuración de la cabecera del PDF
+        self.image('images/EducaciónAprende.jpeg', x=10, y=8, w=35)
+        self.image('images/logo-aprendemx.png', x=46, y=7, w=35)
+        self.ln()
+
+        self.set_font('Arial', 'B', 6)
+        self.cell(323,1, 'SECRETARÍA DE EDUCACIÓN PÚBLICA', 0, 10, 'C')
+        self.ln(3)
+        self.cell(293,1, 'Subdirección de Sistematización de Acervos y Desarrollo Audiovisual', 0, 20, 'C')
+        self.ln(3)
+        self.cell(351.5,1, 'Audiovisual', 0, 20, 'C')
+        self.ln(3)
+        self.cell(306.5,1, 'Departamento de Conservación de Acervos Videográficos', 0, 20, 'C')
+        self.ln(40)
+
+        self.set_font('Arial', 'B', 8)
+        self.cell(84, 10,   'NOMBRE:      _______________________________', 0, 0, 'C')
+        # self.ln(7)
+        self.cell(103, 10,  'DIRECCIÓN:   _______________________________', 0, 0, 'C')
+        self.ln(17)
+        self.cell(83, 10,   'PUESTO:      _______________________________', 0, 0, 'C')
+
+        self.cell(103, 10,  'EXTENSIÓN:   _______________________________', 0, 0, 'C')
+        # self.cell(103, 10,  'CORREO:      _______________________________', 0, 0, 'C')
+        self.ln(40)
+        
+        self.set_font('Arial', 'B', 15)
+        self.cell(180, 10, f'Folio de la cinta ({self.q})', 0, 0, 'C')
+        self.ln(20)
+        
+    def footer(self):
+
+        self.ln(40)
+        self.set_font('Arial', 'B', 8)
+        self.cell(83, 10, 'RECIBE: _______________________________', 0, 0, 'C')
+        self.cell(93, 10, 'DEVUELVE: _______________________________', 0, 0, 'C')
+        self.ln()
+
+        # self.cell(103, 10, '_______________________________', 0, 0, 'C')
+        # self.ln(8)
+        # self.cell(200, 10, '_______________________________', 0, 0, 'C')
+        # self.ln()
+
+        # Configuración del pie de página del PDF
+        self.set_y(-15)
+        self.set_font('Arial', 'I', 8)
+        self.cell(0, 10, 'Página %s' % self.page_no(), 0, 0, 'C')
+
+    def generate_Table(self, data):
+        # Generación del contenido de la tabla en el PDF
+        self.set_fill_color(31, 237, 237)
+        self.set_text_color(6, 28, 28) # Establece el color de la letra en blanco
+        self.cell(90, 10, 'Código de Barras', 1, 0, '', True)
+        self.cell(90, 10, 'Fecha de Devolucón', 1, 0, '', True)
+        self.set_text_color(0, 0, 0)
+        self.ln()
+
+        for row in data:
+            self.cell(90, 10, str(row['vide_codigo']), 1)
+            self.cell(90, 10, str(row['pres_fecha_devolucion']), 1)
+            self.ln()
+
+def generar_pdf_modal(request):
+    q = int(request.GET.get("q"))
+    queryset = DetallePrestamos.objects.filter(pres_folio=q).values('vide_codigo', 'pres_fecha_devolucion')
+
+    response = HttpResponse(content_type='application/pdf')
+    response['Content-Disposition'] = f'attachment; filename="Videoteca_Folio_{q}.pdf"'
+    pdf = GENERATE('P', 'mm', (200, 380), q)
+    pdf.add_page()
+    pdf.generate_Table(queryset)
     
-    input_file = settings.MEDIA_ROOT+ '/Formatos/Compensacion_v1.jrxml'
-    CreateJson(dateInicio, dateFin, None, folio)
-    output_file = settings.MEDIA_ROOT+ '/Formatos'
-    
-    json_query = 'contacts.person'
-
-    
-    #dictionary = {"contacts": {"person": [ {'Folio':34, 'RFC':61, 'Nombre':82, 'ClaveP':82, 'Categoria':82, 'Codigo':82, 'Compensacion':82, 'CostoHora':82, 'HorasXMes':82, 'Importe':82, 'Fecha':82, 'NoFolio':82}  ] } } 
-    #jsonString = json.dumps(dictionary, indent=4)
-
-    #print(jsonString)
-    conn = {
-      'driver': 'json',
-      'data_file': settings.MEDIA_ROOT+ '/Formatos/data.json',
-      'json_query': 'compensacion'
-   }
-    outputFile= settings.MEDIA_ROOT+ '/Formatos/ReporteCompensacion '+fechaInicio+' al '+fechaFin+'.pdf' 
-    pyreportjasper = PyReportJasper()
-    pyreportjasper.config(
-      input_file,
-      output_file=outputFile,
-      output_formats=["pdf"],
-      db_connection=conn
-   )
-    pyreportjasper.process_report()
-    print('Result is the file below.')
-   # output_file = output_file + '.pdf'
-    if os.path.isfile(outputFile):
-    #    print('Report generated successfully!')
-        with open(outputFile, 'rb') as pdf:
-            response = HttpResponse(pdf.read(),content_type='application/pdf')
-            response['Content-Disposition'] = 'filename=ReporteCompensacion '+fechaInicio+' al '+fechaFin+'.pdf'
-        return response
+    response.write(pdf.output(dest='S').encode('latin1'))
+    return response
+   
 
 
-
-def CreateJson(dateInicio, dateFin, condicion, folio):
-    data = {}
-    horas = 0
-    total = 0
-    data['compensacion']=[]
-    if condicion == None:
-        compensaciones = Compensaciones.objects.all().order_by('info_person__rfc')
-    else:
-        compensaciones = Compensaciones.objects.all().filter(Q(info_person__cat_area_org__pk = condicion))
-    if compensaciones.count() > 0:
-        j=0
-        for i, compen in enumerate(compensaciones):
-            print(compen.info_person.apellido1 +" "+ compen.info_person.apellido2 +" "+ compen.info_person.nombres)
-            horas = calculoIncidenciasPersona(dateInicio, dateFin, compen.info_person)
-            horas = math.floor(horas)
-            if horas <= 90:
-                importe = Decimal(horas) * (compen.compensacion.montoUnitario / 90)
-            elif horas == 0:
-                importe = 0
-            else:
-                horas = 90
-                importe = compen.compensacion.montoUnitario 
-            total += importe
-            if horas != 0:
-                data['compensacion'].append({'Folio': j+1,
-                'NoFolio': folio+j,
-                'Area': compen.info_person.cat_area_org.nombre,
-                'RFC': compen.info_person.rfc,
-                'Nombre':  compen.info_person.apellido1.upper() +" "+ compen.info_person.apellido2.upper() +" "+ compen.info_person.nombres.upper() ,
-                'ClaveP':  compen.info_person.clave_presupuestal,
-                'Categoria': compen.compensacion.nombre,
-                'Codigo': compen.compensacion.codigo,
-                'Compensacion': '$ '+ str(compen.compensacion.montoUnitario),
-                'CostoHora': str(round(compen.compensacion.montoUnitario / 90, 4) ),
-                'HorasXMes': str(horas),
-                'Importe': '$ ' +str("{:,}".format(round(importe, 2))),
-                'Fecha': str(dateInicio.strftime("%d-%m-%Y")) + ' al ' +str(dateFin.strftime("%d-%m-%Y")), 
-                'Total': '$ ' + str("{:,}".format(round( total, 2))), 
-                'Logo1': settings.MEDIA_ROOT+ '/Formatos/logo-sep.png', 
-                'Logo2':  settings.MEDIA_ROOT+ '/Formatos/logo-aprendemx.png' })
-                j+=1
-        with open(settings.MEDIA_ROOT+ '/Formatos/data.json', 'w',  encoding='utf8') as file:
-            json.dump(data, file, ensure_ascii=False) 
-    return compensaciones.count()
+# ---------------------------------------------------------------------------------------------------------------------------#
