+++ requirements.txt
@@ -5,3 +5,6 @@ psycopg2==2.9.3
 sqlparse==0.4.2
 tzdata==2021.5
 pyreportjasper
+django-ajax-datatable
+fpdf
+reportlab
